Support for whitelisting.
In some cases you'll find a package that lists as NOT_SPECIFIED or 404_NOT_FOUND, but you know from some other source the license it is actually used as. In that case you can add a `--whitelist` file. The whitelist file can be written as a text file or as a json formatted structure.

The textfile has the following format:
```
<package_name>[: <expected_license>[ -> <map_to_license>]]
```

Valid examples are
```
foo
foobar: NOT_SPECIFIED
barbaz: 404_NOT_FOUND -> MIT
```

The above file will
* whitelist `foo` no matter what license it presents.
* whitelist `foobar` if it presents as NOT_SPECIFIED.
* whitelist `barbaz` if it presents as 404_NOT_FOUND and remap it to MIT.

The remapping is used when listing the output as well as in the `--json` output.

If written as json, the following is eqiuvalent:
```
{
    "foo": {},
    "foobar": {"expected": "NOT_SPECIFIED"},
    "barbaz": {"expected": ""404_NOT_FOUND", "mapto": "MIT"}
}
```