Added support to generate credits-file.
It is also possible to generate an output file straight from pkglic, by means of `--credits` and `--creditstemplate`. The credits template will be expanded by jinja2, with a list of packages as described above passed in named `packages`.

There is a simple template included as default if you omit the `--creditstemplate` argument.